Extarcting main packages from both carbonapi and carbonzipper, into their separate package. The resulting main packages are quite rudimentary, just reading the configs. This also makes sure that we need to do very little to make any booking related config changes, and prevents the pain of having to keep the main packages in sync across our public and private github repos, simplifying the release process.
