+++ app/carbonapi/http_handlers_test.go
@@ -1,4 +1,4 @@
-package main
+package carbonapi
 
 import (
 	"net/http"
